formupdated

+++ src/services/apiCabins.ts
@@ -1,43 +1,55 @@
-import supabase from "./superbase";
+import supabase, { supabaseUrl } from "./superbase.ts";
 export async function getCabins() {
-    const {data,error}=await supabase.from('cabins').select('*')
-    if(error) throw new Error(error.message)
-    return data
+  const { data, error } = await supabase.from('cabins').select('*')
+  if (error) throw new Error(error.message)
+  return data
 }
 
-export async function deleteCabin(id:string){
- const { data,error } = await supabase
-        .from('cabins')
-        .delete()
-        .eq('id', id)
-    if(error) throw new Error(error.message)
-    console.log(data,'data')
-    return data
+export async function deleteCabin(id: string) {
+  const { data, error } = await supabase
+    .from('cabins')
+    .delete()
+    .eq('id', id)
+  if (error) throw new Error(error.message)
+  console.log(data, 'data')
+  return data
 }
 
-export async function createNewCabin(newCabin:any) {
+export async function createEditCabin(newCabin, id) {
+  const hasImagePath = newCabin.image?.startsWith?.(supabaseUrl);
+
   const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
     "/",
     ""
-    );
-    console.log(newCabin,'newCabin')
-    // https://lfvrvjnnobpkhjbgprup.supabase.co/storage/v1/object/public/cabin-images/cabin-001.jpg
-    const supabaseUrl='https://lfvrvjnnobpkhjbgprup.supabase.co'
-     const imagePath = 
-   `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-    const { data, error } = await supabase
-    .from('cabins')
-    .insert([
-        { ...newCabin, image: imagePath }
-    ])
-  .select()
-    if (error) throw new Error(error.message)
-    
-    // 2. Upload image
+  );
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+
+  // A) CREATE
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+
+  // B) EDIT
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
+
+  if (error) {
+    console.error(error);
+    throw new Error("Cabin could not be created");
+  }
+
+  // 2. Upload image
+  if (hasImagePath) return data;
+
   const { error: storageError } = await supabase.storage
     .from("cabin-images")
     .upload(imageName, newCabin.image);
-     // 3. Delete the cabin IF there was an error uplaoding image
+
+  // 3. Delete the cabin IF there was an error uplaoding image
   if (storageError) {
     await supabase.from("cabins").delete().eq("id", data.id);
     console.error(storageError);
@@ -45,5 +57,6 @@ export async function createNewCabin(newCabin:any) {
       "Cabin image could not be uploaded and the cabin was not created"
     );
   }
-    return data
+
+  return data;
 }
